Rename 'actions' to 'elements' (matches Figma)
